project/email_application/requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
Django==2.2.4
django-appconf==1.0.3
django-celery==3.3.1
django-celery-email==2.0.2
djangorestframework==3.10.2
importlib-metadata==0.19
kombu==4.6.4
more-itertools==7.2.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
